requirements.txt
fastapi
uvicorn
pymongo
python-jose
passlib>=1.7.4
bcrypt==3.2.0  # Downgrade bcrypt to a compatible version
pyjwt